beatsynthconverter use internal classes
beatsynthconverter updated to use the internal classes and bugfix for converting audio file that already exists

+++ Assets/MikuEditor/Scripts/Nova_Max/BeatsynthConverter.cs
@@ -4,30 +4,28 @@ using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using UnityEngine;
+using MiKu.NET.Charting;
 
-public static class BeatsynthConverter
+public static class BeatSynthConverter
 {
     private static float sqrttwo = (float)Math.Sqrt(2);
 
-    public static string Convert(string path, int distancex = 50, int distancey = 35, int offsety = 10, int dynamic = 20, int repeat = 200)
+    public static Chart Convert(string path, int distancex = 50, int distancey = 35, int offsety = 10, int dynamic = 20, int repeat = 200)
     {
-        string result="";
         path = path.Substring(0, path.LastIndexOf(@"\"));
 
-        if (File.Exists(path + "\\info.dat"))
+        Chart synthSong = new Chart();
+
+        if (File.Exists(path + @"\info.dat"))
         {
             //load song info
             string infoStr = "";
-            using (StreamReader streamReader = new StreamReader(path + "\\info.dat"))
+            using (StreamReader streamReader = new StreamReader(path + @"\info.dat"))
             {
                 infoStr = streamReader.ReadToEnd();
             }
             //deserialize song info
             Info info = JsonConvert.DeserializeObject<Info>(infoStr);
-
-            //initialize SynthSong with parameters
-            SynthSong synthSong = new SynthSong();
-
 
             synthSong.Name = (info._songName != "") ? info._songName : "N/A";
             synthSong.Author = (info._songAuthorName != "") ? info._songAuthorName : "N/A";
@@ -35,13 +33,22 @@ public static class BeatsynthConverter
             synthSong.Offset = (info._songTimeOffset < 0) ? info._songTimeOffset : synthSong.Offset = 0;
             synthSong.Beatmapper = (info._levelAuthorName != "") ? info._levelAuthorName : "N/A";
 
+            synthSong.Artwork = "Default Artwork";
+            synthSong.ArtworkBytes = null;
+            synthSong.AudioName = "";
+            synthSong.AudioData = null;
+            synthSong.AudioFrecuency = 0;
+            synthSong.AudioChannels = 2;
+            synthSong.EditorVersion = "1.8";
+            synthSong.IsAdminOnly = false;
+
             for (int i = 0; i < info._difficultyBeatmapSets[0]._difficultyBeatmaps.Count && i < 5; i++)
             {
-                if (File.Exists(path + "\\" + info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._beatmapFilename))
+                if (File.Exists(path + @"\" + info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._beatmapFilename))
                 {
                     //load song difficulty
                     string beatSongStr = "";
-                    using (StreamReader streamReader = new StreamReader(path + "\\" + info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._beatmapFilename))
+                    using (StreamReader streamReader = new StreamReader(path + @"\" + info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._beatmapFilename))
                     {
                         beatSongStr = streamReader.ReadToEnd();
                     }
@@ -49,80 +56,62 @@ public static class BeatsynthConverter
                     //deserialize song difficulty
                     BeatSong beatSong = JsonConvert.DeserializeObject<BeatSong>(beatSongStr);
 
-                    string difficulty = "Master";
-
-                    if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Expert"))
-                    {
-                        difficulty = "Expert";
-                    }
-                    else if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Hard"))
-                    {
-                        difficulty = "Hard";
-                    }
-                    else if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Normal"))
-                    {
-                        difficulty = "Normal";
-                    }
-                    else if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Easy"))
-                    {
-                        difficulty = "Easy";
-                    }
-
                     //convert notes
+                    Dictionary<float, List<Note>> track = new Dictionary<float, List<Note>>();
                     for (int j = 0; j < beatSong._notes.Count; j++)
                     {
-                        SynthNote note = new SynthNote();
-                        note.Id = i.ToString() + "," + j.ToString();
-
                         int time = (int)Math.Round(beatSong._notes[j]._time * 60000f / info._beatsPerMinute);
 
-                        float x = (distancex / 150f) * (beatSong._notes[j]._lineIndex - 1.5f);
-                        float y = (distancey / 100f) * (beatSong._notes[j]._lineLayer - 1f) + offsety * 0.01f;
-                        float z = (time) * 0.02f;
-
+                        float y_offset = 0;
+                        float _offset = 0;
                         switch (beatSong._notes[j]._cutDirection)
                         {
-                            case 0: y += dynamic * 0.01f; break;//up
-                            case 1: y -= dynamic * 0.01f; break;//down
-                            case 2: x += dynamic * 0.01f; break;//left
-                            case 3: x -= dynamic * 0.01f; break;//right
-                            case 4: y += (dynamic / sqrttwo) * 0.01f; x += (dynamic / sqrttwo) * 0.01f; break;//up_left
-                            case 5: y += (dynamic / sqrttwo) * 0.01f; x -= (dynamic / sqrttwo) * 0.01f; break;//up_right
-                            case 6: y -= (dynamic / sqrttwo) * 0.01f; x += (dynamic / sqrttwo) * 0.01f; break;//down_left
-                            case 7: y -= (dynamic / sqrttwo) * 0.01f; x -= (dynamic / sqrttwo) * 0.01f; break;//down_right
+                            case 0: y_offset += dynamic * 0.01f; break;//up
+                            case 1: y_offset -= dynamic * 0.01f; break;//down
+                            case 2: _offset += dynamic * 0.01f; break;//left
+                            case 3: _offset -= dynamic * 0.01f; break;//right
+                            case 4: y_offset += (dynamic / sqrttwo) * 0.01f; _offset += (dynamic / sqrttwo) * 0.01f; break;//up_left
+                            case 5: y_offset += (dynamic / sqrttwo) * 0.01f; _offset -= (dynamic / sqrttwo) * 0.01f; break;//up_right
+                            case 6: y_offset -= (dynamic / sqrttwo) * 0.01f; _offset += (dynamic / sqrttwo) * 0.01f; break;//down_left
+                            case 7: y_offset -= (dynamic / sqrttwo) * 0.01f; _offset -= (dynamic / sqrttwo) * 0.01f; break;//down_right
                         }
 
-                        note.Position = new float[3] { x, y, z };
+                        Vector3 pos = new Vector3(_offset + (distancex / 150f) * (beatSong._notes[j]._lineIndex - 1.5f), y_offset + (distancey / 100f) * (beatSong._notes[j]._lineLayer - 1f) + offsety * 0.01f, (time) * 0.02f);
 
+                        Note.NoteType type;
                         if (beatSong._notes[j]._type == 0)
                         {
-                            note.Type = 1;
+                            type = Note.NoteType.LeftHanded;
                         }
                         else if (beatSong._notes[j]._type == 1)
                         {
-                            note.Type = 0;
+                            type = Note.NoteType.RightHanded;
                         }
                         else//bombs
                         {
                             continue;
                         }
 
-                        string key = time.ToString();
+                        string id = i.ToString() + "," + j.ToString();
 
-                        if (!synthSong.Track[difficulty].ContainsKey(key))
+                        Note note = new Note(pos, id,-1, type);
+
+                        if (!track.ContainsKey(time))
                         {
-                            List<SynthNote> notes = new List<SynthNote>();
+                            List<Note> notes = new List<Note>();
                             notes.Add(note);
 
-                            synthSong.Track[difficulty].Add(key, notes);
+                            track.Add(time, notes);
                         }
                         else
                         {
-                            synthSong.Track[difficulty][key].Add(note);
+                            track[time].Add(note);
                         }
                     }
 
                     //convert obstacles
+                    List<float> crouches = new List<float>();
+                    List<Slide> slides = new List<Slide>();
                     for (int j = 0; j < beatSong._obstacles.Count; j++)
                     {
                         int time = (int)Math.Round(beatSong._obstacles[j]._time * 60000f / info._beatsPerMinute);
@@ -132,54 +121,84 @@ public static class BeatsynthConverter
                         {
                             if (beatSong._obstacles[j]._type == 1)
                             {
-                                synthSong.Crouchs[difficulty].Add(n);
+                                crouches.Add(n);
                             }
                             else
                             {
-                                CSlide slide = new CSlide();
+                                Slide slide = new Slide();
                                 slide.time = n;
                                 if (beatSong._obstacles[j]._width == 1)
                                 {
                                     if (beatSong._obstacles[j]._lineIndex == 0)//one on the left
                                     {
-                                        slide.SlideType = 4;
+                                        slide.slideType = (Note.NoteType)4;
                                     }
                                     else if (beatSong._obstacles[j]._lineIndex == 3)//one on the right
                                     {
-                                        slide.SlideType = 2;
+                                        slide.slideType = (Note.NoteType)2;
                                     }
                                     else//one middle right or middle left
                                     {
-                                        slide.SlideType = 3;
+                                        slide.slideType = (Note.NoteType)3;
                                     }
-                                    synthSong.Slides[difficulty].Add(slide);
+                                    slides.Add(slide);
                                 }
                                 else if (beatSong._obstacles[j]._width == 2)
                                 {
                                     if (beatSong._obstacles[j]._lineIndex == 0)//two on the left
                                     {
-                                        slide.SlideType = 1;
+                                        slide.slideType = (Note.NoteType)1;
                                     }
                                     else if (beatSong._obstacles[j]._lineIndex == 1)//two in the middle
                                     {
-                                        slide.SlideType = 3;
+                                        slide.slideType = (Note.NoteType)3;
                                     }
                                     else if (beatSong._obstacles[j]._lineIndex == 2)//two on the right
                                     {
-                                        slide.SlideType = 0;
+                                        slide.slideType = (Note.NoteType)0;
                                     }
-                                    synthSong.Slides[difficulty].Add(slide);
+                                    slides.Add(slide);
                                 }
                                 else
                                 {
-                                    synthSong.Crouchs[difficulty].Add(n);
+                                    crouches.Add(n);
                                 }
                             }
                         }
                     }
+
+                    if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Expert"))
+                    {
+                        synthSong.Track.Expert = track;
+                        synthSong.Crouchs.Expert = crouches;
+                        synthSong.Slides.Expert = slides;
+                    }
+                    else if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Hard"))
+                    {
+                        synthSong.Track.Hard = track;
+                        synthSong.Crouchs.Hard = crouches;
+                        synthSong.Slides.Hard = slides;
+                    }
+                    else if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Normal"))
+                    {
+                        synthSong.Track.Normal = track;
+                        synthSong.Crouchs.Normal = crouches;
+                        synthSong.Slides.Normal = slides;
+                    }
+                    else if (info._difficultyBeatmapSets[0]._difficultyBeatmaps[i]._difficulty.Equals("Easy"))
+                    {
+                        synthSong.Track.Easy = track;
+                        synthSong.Crouchs.Easy = crouches;
+                        synthSong.Slides.Easy = slides;
+                    }
+                    else
+                    {
+                        synthSong.Track.Master = track;
+                        synthSong.Crouchs.Master = crouches;
+                        synthSong.Slides.Master = slides;
+                    }
                 }
             }
-
 
             string[] filePaths = Directory.GetFiles(path);
 
@@ -188,12 +207,13 @@ public static class BeatsynthConverter
                 if (filePath.Contains(".egg"))//replace the .egg file ending with .ogg
                 {
                     string tmp = filePath.Substring(0, filePath.IndexOf(".egg")) + ".ogg";
-                    File.Copy(filePath, tmp);
+                    if (!File.Exists(tmp))
+                    {
+                        File.Copy(filePath, tmp);
+                    }
                 }
             }
-            //serialize SynthSong
-           result = JsonConvert.SerializeObject(synthSong);  
         }
-        return result;
+        return synthSong;
     }
 }
